Image classification model on cropped dataset
The dataset is first filtered to only be images of one object, then if there's bounding boxes the image is cropped to the bounding boxes, then saved into a folder named after the class they belong in. The model is then run on the folders doing a 80/20 training/test split.